Poprawa funkcjonalności tworzenia scenariuszy. Wstępna analiza i przygotowanie do generowania JAVADOC.
