--- src/main/java/pl/put/poznan/restservice/logic/ElseStep.java
@@ -1,7 +1,0 @@
-package pl.put.poznan.restservice.logic;
-
-public record ElseStep(Step[] scenario) implements ScenarioElement {
-    public void accept(ScenarioVisitor scenarioVisitor) {
-        scenarioVisitor.visit(this);
-    }
-}
