+++ src/main/java/pl/put/poznan/restservice/logic/IfStep.java
@@ -1,6 +1,6 @@
 package pl.put.poznan.restservice.logic;
 
-public record IfStep(String condition, Step[] scenario, ElseStep elseStep) implements ScenarioElement  {
+public record IfStep(String condition, Step[] scenario, Step[] elseStep) implements ScenarioElement  {
     public void accept(ScenarioVisitor scenarioVisitor) {
         scenarioVisitor.visit(this);
     }
