+++ src/main/java/pl/put/poznan/restservice/logic/ScenarioVisitor.java
@@ -4,6 +4,5 @@ public interface ScenarioVisitor{
     void visit(Scenario scenario);
     void visit(Step step);
     void visit(IfStep ifStep);
-    void visit(ElseStep elseStep);
     void visit(ForEachStep forEachStep);
 }
