+++ src/main/java/pl/put/poznan/restservice/logic/Step.java
@@ -1,6 +1,6 @@
 package pl.put.poznan.restservice.logic;
 
-public record Step(String step, IfStep ifStep, ElseStep elseStep, ForEachStep forEachStep) implements ScenarioElement  {
+public record Step(String step, IfStep ifStep, ForEachStep forEachStep) implements ScenarioElement  {
     public void accept(ScenarioVisitor scenarioVisitor) {
         scenarioVisitor.visit(this);
     }
